Using a common ecto_enum_test for pg and mysql
