+++ test/mysql/test_helper.exs
@@ -1,22 +1,25 @@
+Logger.configure(level: :info)
 ExUnit.start()
+alias Ecto.Integration.TestRepo
+alias Ecto.Integration.PoolRepo
 
 Code.require_file "ecto_migration.exs", __DIR__
 
-Application.put_env(:ecto, Ecto.Integration.TestRepo,
+Application.put_env(:ecto, TestRepo,
   adapter: Ecto.Adapters.MySQL,
   url: "ecto://root@localhost/ecto_test",
   pool: Ecto.Adapters.SQL.Sandbox)
 
-defmodule Ecto.Integration.TestRepo do
+defmodule TestRepo do
   use Ecto.Repo, otp_app: :ecto
 end
 
-Application.put_env(:ecto, Ecto.Integration.PoolRepo,
+Application.put_env(:ecto, PoolRepo,
   adapter: Ecto.Adapters.MySQL,
   url: "ecto://root@localhost/ecto_test",
   pool_size: 10)
 
-defmodule Ecto.Integration.PoolRepo do
+defmodule PoolRepo do
   use Ecto.Repo, otp_app: :ecto
 
   def create_prefix(prefix) do
@@ -37,9 +40,6 @@ defmodule Ecto.Integration.Case do
   end
 end
 
-alias Ecto.Integration.TestRepo
-alias Ecto.Integration.PoolRepo
-
 {:ok, _} = Ecto.Adapters.MySQL.ensure_all_started(TestRepo, :temporary)
 
 # Load up the repository, start it, and run migrations
@@ -47,8 +47,8 @@ _   = Ecto.Adapters.MySQL.storage_down(TestRepo.config)
 :ok = Ecto.Adapters.MySQL.storage_up(TestRepo.config)
 
 {:ok, _pid} = TestRepo.start_link
-{:ok, _pid} = PoolRepo.start_link
+# {:ok, _pid} = PoolRepo.start_link
 :ok = Ecto.Migrator.up(TestRepo, 0, Ecto.Integration.Migration, log: false)
-Ecto.Adapters.SQL.Sandbox.mode(TestRepo, :manual)
+Ecto.Adapters.SQL.Sandbox.mode(TestRepo, :auto)
 Process.flag(:trap_exit, true)
